+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/client/model/prisoner/PrisonerAddress.kt
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty
 import java.time.LocalDate
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-data class Address(
+data class PrisonerAddress(
   @JsonProperty("postalCode")
   val postcode: String? = null,
   val fullAddress: String? = null,
