+++ src/main/kotlin/uk/gov/justice/digital/hmpps/personrecord/model/person/Address.kt
@@ -2,6 +2,7 @@ package uk.gov.justice.digital.hmpps.personrecord.model.person
 
 import uk.gov.justice.digital.hmpps.personrecord.api.model.probation.ProbationCreateAddress
 import uk.gov.justice.digital.hmpps.personrecord.client.model.offender.ProbationAddress
+import uk.gov.justice.digital.hmpps.personrecord.client.model.prisoner.PrisonerAddress
 import uk.gov.justice.digital.hmpps.personrecord.client.model.sas.SasAddressData
 import uk.gov.justice.digital.hmpps.personrecord.extensions.nullIfBlank
 import uk.gov.justice.digital.hmpps.personrecord.extensions.toUkZonedDateTime
@@ -18,7 +19,6 @@ import kotlin.reflect.full.memberProperties
 import uk.gov.justice.digital.hmpps.personrecord.api.model.sysconsync.PrisonAddress as SysconAddress
 import uk.gov.justice.digital.hmpps.personrecord.client.model.court.commonplatform.Address as CommonPlatformAddress
 import uk.gov.justice.digital.hmpps.personrecord.client.model.court.libra.Address as LibraAddress
-import uk.gov.justice.digital.hmpps.personrecord.client.model.prisoner.Address as PrisonerAddress
 
 data class Address(
   val noFixedAbode: Boolean? = null,
